// project-3-front-end-beta-main/src/app/models/user.model.ts
export class User {

    user_id: number = 0;
    email: string = "";
    username: string = "";
    password: string = "";
    first_name: string = "";
    last_name: string = "";
    address: string = "";
    userType: string = "";
    contact: string = "";
    imageUrl: string = "";
    userRemoved: boolean = false;
}